smartremove tests in own file
