+++ common/test/test_snapshots.py
@@ -300,146 +300,6 @@ class Callbacks(generic.SnapshotsTestCase):
                              '[E] Error: rsync: send_files failed to open "/foo/bar": Operation not permitted (1)\n', f.read())
 
 
-class SmartRemove(generic.SnapshotsTestCase):
-    def test_increment_month(self):
-        self.assertEqual(self.sn.incMonth(date(2016,  4, 21)), date(2016, 5, 1))
-        self.assertEqual(self.sn.incMonth(date(2016, 12, 24)), date(2017, 1, 1))
-
-    def test_decrement_month(self):
-        self.assertEqual(self.sn.decMonth(date(2016, 4, 21)), date(2016,  3, 1))
-        self.assertEqual(self.sn.decMonth(date(2016, 1, 14)), date(2015, 12, 1))
-
-    def test_keep_all(self):
-        sid1 = snapshots.SID('20160424-215134-123', self.cfg)
-        sid2 = snapshots.SID('20160422-030324-123', self.cfg)
-        sid3 = snapshots.SID('20160422-020324-123', self.cfg)
-        sid4 = snapshots.SID('20160422-010324-123', self.cfg)
-        sid5 = snapshots.SID('20160421-013218-123', self.cfg)
-        sid6 = snapshots.SID('20160410-134327-123', self.cfg)
-        sids = [sid1, sid2, sid3, sid4, sid5, sid6]
-
-        keep = self.sn.smartRemoveKeepAll(sids,
-                                               date(2016, 4, 20),
-                                               date(2016, 4, 23))
-        self.assertSetEqual(keep, set((sid2, sid3, sid4, sid5)))
-
-        keep = self.sn.smartRemoveKeepAll(sids,
-                                               date(2016, 4, 11),
-                                               date(2016, 4, 18))
-        self.assertSetEqual(keep, set())
-
-    def test_keep_first(self):
-        sid1 = snapshots.SID('20160424-215134-123', self.cfg)
-        sid2 = snapshots.SID('20160422-030324-123', self.cfg)
-        sid3 = snapshots.SID('20160422-020324-123', self.cfg)
-        sid4 = snapshots.SID('20160422-010324-123', self.cfg)
-        sid5 = snapshots.SID('20160421-013218-123', self.cfg)
-        sid6 = snapshots.SID('20160410-134327-123', self.cfg)
-        sids = [sid1, sid2, sid3, sid4, sid5, sid6]
-
-        keep = self.sn.smartRemoveKeepFirst(sids,
-                                            date(2016, 4, 20),
-                                            date(2016, 4, 23))
-        self.assertSetEqual(keep, set((sid2,)))
-
-        keep = self.sn.smartRemoveKeepFirst(sids,
-                                            date(2016, 4, 11),
-                                            date(2016, 4, 18))
-        self.assertSetEqual(keep, set())
-
-    def test_keep_first_no_errors(self):
-        sid1 = snapshots.SID('20160424-215134-123', self.cfg)
-        sid2 = snapshots.SID('20160422-030324-123', self.cfg)
-        sid2.makeDirs()
-        sid2.failed = True
-        sid3 = snapshots.SID('20160422-020324-123', self.cfg)
-        sid4 = snapshots.SID('20160422-010324-123', self.cfg)
-        sid5 = snapshots.SID('20160421-013218-123', self.cfg)
-        sid6 = snapshots.SID('20160410-134327-123', self.cfg)
-        sids = [sid1, sid2, sid3, sid4, sid5, sid6]
-
-        # keep the first healthy snapshot
-        keep = self.sn.smartRemoveKeepFirst(sids,
-                                            date(2016, 4, 20),
-                                            date(2016, 4, 23),
-                                            keep_healthy = True)
-        self.assertSetEqual(keep, set((sid3,)))
-
-        # if all snapshots failed, keep the first at all
-        for sid in (sid3, sid4, sid5):
-            sid.makeDirs()
-            sid.failed = True
-        keep = self.sn.smartRemoveKeepFirst(sids,
-                                            date(2016, 4, 20),
-                                            date(2016, 4, 23),
-                                            keep_healthy = True)
-        self.assertSetEqual(keep, set((sid2,)))
-
-    def test_smart_remove_list(self):
-        sid1  = snapshots.SID('20160424-215134-123', self.cfg)
-        sid2  = snapshots.SID('20160422-030324-123', self.cfg)
-        sid3  = snapshots.SID('20160422-020324-123', self.cfg)
-        sid4  = snapshots.SID('20160422-010324-123', self.cfg)
-        sid5  = snapshots.SID('20160421-033218-123', self.cfg)
-        sid6  = snapshots.SID('20160421-013218-123', self.cfg)
-        sid7  = snapshots.SID('20160420-013218-123', self.cfg)
-        sid8  = snapshots.SID('20160419-013218-123', self.cfg)
-        sid9  = snapshots.SID('20160419-003218-123', self.cfg)
-        sid10 = snapshots.SID('20160418-003218-123', self.cfg)
-        sid11 = snapshots.SID('20160417-033218-123', self.cfg)
-        sid12 = snapshots.SID('20160417-003218-123', self.cfg)
-        sid13 = snapshots.SID('20160416-134327-123', self.cfg)
-        sid14 = snapshots.SID('20160416-114327-123', self.cfg)
-        sid15 = snapshots.SID('20160415-134327-123', self.cfg)
-        sid16 = snapshots.SID('20160411-134327-123', self.cfg)
-        sid17 = snapshots.SID('20160410-134327-123', self.cfg)
-        sid18 = snapshots.SID('20160409-134327-123', self.cfg)
-        sid19 = snapshots.SID('20160407-134327-123', self.cfg)
-        sid20 = snapshots.SID('20160403-134327-123', self.cfg)
-        sid21 = snapshots.SID('20160402-134327-123', self.cfg)
-        sid22 = snapshots.SID('20160401-134327-123', self.cfg)
-        sid23 = snapshots.SID('20160331-134327-123', self.cfg)
-        sid24 = snapshots.SID('20160330-134327-123', self.cfg)
-        sid25 = snapshots.SID('20160323-133715-123', self.cfg)
-        sid26 = snapshots.SID('20160214-134327-123', self.cfg)
-        sid27 = snapshots.SID('20160205-134327-123', self.cfg)
-        sid28 = snapshots.SID('20160109-134327-123', self.cfg)
-        sid29 = snapshots.SID('20151224-134327-123', self.cfg)
-        sid30 = snapshots.SID('20150904-134327-123', self.cfg)
-        sid31 = snapshots.SID('20140904-134327-123', self.cfg)
-
-        sids = [       sid1,  sid2,  sid3,  sid4,  sid5,  sid6,  sid7,  sid8,  sid9,
-                sid10, sid11, sid12, sid13, sid14, sid15, sid16, sid17, sid18, sid19,
-                sid20, sid21, sid22, sid23, sid24, sid25, sid26, sid27, sid28, sid29,
-                sid30, sid31]
-        for sid in sids:
-            sid.makeDirs()
-        now = datetime(2016, 4, 24, 21, 51, 34)
-
-        del_snapshots = self.sn.smartRemoveList(now,
-                                                3, #keep_all
-                                                7, #keep_one_per_day
-                                                5, #keep_one_per_week
-                                                3  #keep_one_per_month
-                                                )
-        self.assertListEqual(del_snapshots, [sid6, sid9, sid12, sid13, sid14,
-                                             sid15, sid16, sid18, sid19, sid21,
-                                             sid22, sid24, sid27, sid28, sid30])
-
-        # test failed snapshots
-        for sid in (sid5, sid8, sid11, sid12, sid20, sid21, sid22):
-            sid.failed = True
-        del_snapshots = self.sn.smartRemoveList(now,
-                                                3, #keep_all
-                                                7, #keep_one_per_day
-                                                5, #keep_one_per_week
-                                                3  #keep_one_per_month
-                                                )
-        self.assertListEqual(del_snapshots, [sid5, sid8, sid11, sid12, sid14,
-                                             sid15, sid16, sid18, sid19, sid20, sid21,
-                                             sid22, sid24, sid27, sid28, sid30])
-
-
 class SnapshotWithSID(generic.SnapshotsWithSidTestCase):
     def test_backup_config(self):
         self.sn.backupConfig(self.sid)
